content-security-policy/reporting/report-clips-sample.html

<!DOCTYPE html>
<html>
<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/content-security-policy/support/testharness-helper.js"></script>
  <meta http-equiv="Content-Security-Policy" content="script-src 'self' 'nonce-abc' 'report-sample'; trusted-types default">
</head>
<body>
  <script nonce="abc">
  TrustedTypes.createPolicy("default", {createScript: s => s, createHTML: s => null});

  async_test(t => {
    waitUntilCSPEventForEval(t, 21).then(t.step_func_done(e => {
      assert_equals(e.sample, "evil = '12345678901234567890123456789012");
      assert_less_than_equal(e.sample.length, 40);
    }));

    let evil = false;
    assert_throws(new EvalError(), _ => {
      eval("evil = '1234567890123456789012345678901234567890';");
    });
    assert_false(evil);
  }, "Unsafe eval violation sample is clipped to 40 characters.");

  async_test(t => {
    waitUntilCSPEventForTrustedTypes(t, 34).then(t.step_func_done(e => {
      assert_equals(e.sample, "Element.innerHTML 1234567890123456789012345678901234567890");
      assert_less_than_equal(e.sample.length, 40 + "Element.innerHTML ".length);
    }));

    const a = document.createElement("a");
    assert_throws(new TypeError(), _ => {
      a.innerHTML = "1234567890123456789012345678901234567890xxxx";
    });
    assert_equals(a.innerHTML, "");
  }, "Trusted Types violation sample is clipped to 40 characters excluded the sink name.");
  </script>
</body>
</html>
